+++ nes-nautilus/src/Nautilus/Interface/BufferRef/ColumnTupleBufferRef.cpp
@@ -56,7 +56,7 @@ Record ColumnTupleBufferRef::readRecord(
 {
     Record record;
     const auto bufferAddress = recordBuffer.getMemArea();
-    for (nautilus::static_val<uint64_t> i = 0; i < fields.size(); ++i)
+    for (nautilus::static_val<uint64_t> i = 0; i < static_cast<uint64_t>(fields.size()); ++i)
     {
         const auto& [name, type, columnOffset] = fields.at(i);
         if (not includesField(projections, name))
@@ -77,7 +77,7 @@ void ColumnTupleBufferRef::writeRecord(
     const nautilus::val<AbstractBufferProvider*>& bufferProvider) const
 {
     const auto bufferAddress = recordBuffer.getMemArea();
-    for (nautilus::static_val<uint64_t> i = 0; i < fields.size(); ++i)
+    for (nautilus::static_val<uint64_t> i = 0; i < static_cast<uint64_t>(fields.size()); ++i)
     {
         const auto& [name, type, columnOffset] = fields.at(i);
         if (not rec.hasField(name))
