+++ nes-physical-operators/src/Aggregation/AggregationBuildPhysicalOperator.cpp
@@ -133,7 +133,7 @@ void AggregationBuildPhysicalOperator::execute(ExecutionContext& ctx, Record& re
         hashMapPtr, hashMapOptions.fieldKeys, hashMapOptions.fieldValues, hashMapOptions.entriesPerPage, hashMapOptions.entrySize);
 
     /// Calling the key functions to add/update the keys to the record
-    for (nautilus::static_val<uint64_t> i = 0; i < hashMapOptions.fieldKeys.size(); ++i)
+    for (nautilus::static_val<uint64_t> i = 0; i < static_cast<uint64_t>(hashMapOptions.fieldKeys.size()); ++i)
     {
         const auto& [fieldIdentifier, type, fieldOffset] = hashMapOptions.fieldKeys[i];
         const auto& function = hashMapOptions.keyFunctions[i];
